requirements.txt

streamlit
torch
opencv-python
Pillow
numpy
